requirements.txt
psycopg[binary]
boto3
python-dotenv